Extract component behavior from GameObject into new Entity class
